refactor: split debug symbols workflow modules
